Rename create_hazard.py script to utility_tutorial.py and add funciton to reassing dates for strom tc data sets from the climada api
